Create Readme.md
### Applications of RNN

One to One: given some scores of a championship, you can predict the winner. 

One to Many: given an image, you can predict what the caption is going to be.

Many to One: given a tweet, you can predict the sentiment of that tweet. 

Many to Many: given an English sentence, you can translate it to its German equivalent.